[DRAFT] Scoped content adapter
ScopedContentAdapter behaves like a normal EMF content adapter, but only
listen to changes in an explicit set of types. It relies on a static
analysis of all the types it can encounter to avoid installing itself
greedily in parts of the model where it can safely determine that no
instance of the types in its scope can be found.

Change-Id: I71d96ed8f2ec4118eca60075b128749990c2b887