[DRAFT 456410] Hide most uses of ECrossReferenceAdapter behind new InverseReferenceFinder interface
The new InverseReferenceFinder interface exposes the sub-set of the
ECrossReferenceAdapter API which is actually used in Sirius, but does
not inherit from ECrossReferenceAdapter or any other related types, and
thus does not enforce a particular implementation.

A new IInterpreter2 interface is also introduced to avoid breaking
existing implementations (e.g. the legacy interpreter) which expect an
ECrossReferenceAdapter and do not know about InverseReferenceFinder.

Bug: 456410
Change-Id: Iaf9f923604fe61910f5009daecb80b1ca989c25f
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>