[DRAFT 485650] Ensure user-defined colors appear first in combo
If the specifier took the time to declare custom colors in the VSM, make
them easier to access by putting them at the top of the combo choice,
instead of after the long list of predefined ones.

Mention this change in the release notes, as otherwise specifiers who
are used to look for their colors at the end can be confused not to find
them there anymore.

Bug: 485650
Change-Id: Icfe269d8079b45ee0bb72beab56552c079ab1f7c
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>