[537507] Move pure-domain code into new package
Start to isolate the pure domain/business code from the HTTP-specific
services, which should become simple glue.

Also rename WorkflowHelper into simply Workflow: it will become the
main entry point for the runtime API to deal with the workflow
associated to a session. The VSM element are named
WorkflowDescription, so there is no ambiguity in calling this one
Workflow.

Bug: 537507
Change-Id: Ie5401c5ea82a43c3357ba05e6e5073d7bb3cd03d
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>