[542859] Lift most evaluateXXX() as default implementations in the interface
Now that the core evaluation if performed in the plain
evaluate(EObject, String) method for almost all interpreters, all the
variants have the same form in most interpreters: evaluate + convert.

Add IInterpreter.getConverter() so that default implementations which
implement this pattern can be defined directly in the interface, and
remove all the overrides that this makes obsolete.

Note that it is possible that IInterpreter.getConverter() is only
needed during the transition to the final form of the API, and may be
removed later in the patch series.

This introduce a small change in the rules used for the Acceleo/MTL
interpreter when using evaluateInteger(): if the expression returned
something which could not be interpreted as an Integer, the method
used to return null. Now it returns 0 (Integer.valueOf(0) to be
precise), which corresponds to the behavior of AQL.

Bug: 542859
Change-Id: I6e5cd583b3e67c435f9cf3639bd6459ce8d1dfa8
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>